move shared user deletion listener separately
